requirements.txt
blinker==1.6.2
click==8.1.3
dominate==2.7.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
visitor==0.1.3
Werkzeug==2.3.3
